backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
mangum==0.17.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0
bcrypt==4.2.0
PyJWT==2.9.0
python-multipart==0.0.12

scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.3
